app/requirements.txt
boto3==1.21.0
python-jenkins==1.7.0
requests
Flask==1.1.4
MarkupSafe==1.1.1
mariadb==1.0.9
PyJWT==2.3.0